refactor: ChessGame을 Controller로 분리
