+++ src/main/java/chess/domain/board/service/BoardService.java
@@ -8,8 +8,13 @@ import chess.domain.board.state.GameProgressState;
 import chess.domain.board.state.StateName;
 import chess.domain.game.dao.GameRepository;
 import chess.domain.piece.Piece;
+import chess.domain.square.File;
+import chess.domain.square.Rank;
 import chess.domain.square.Square;
+import chess.view.PieceName;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 public class BoardService {
@@ -49,5 +54,30 @@ public class BoardService {
         if (gameProgressState.getSateName().equals(StateName.GAME_OVER)) {
             gameRepository.update(gameId, gameProgressState.findWinner());
         }
+    }
+
+    public List<String> makeBoardOutput(Board board) {
+        Map<Square, Piece> boardOutput = board.toBoardOutput().board();
+        List<String> output = new ArrayList<>();
+
+        for (Rank rank : Rank.reverse()) {
+            output.add(makeRankOutput(rank, boardOutput));
+        }
+
+        return output;
+    }
+
+    private String makeRankOutput(Rank rank, Map<Square, Piece> boardOutput) {
+        StringBuilder stringBuilder = new StringBuilder();
+
+        for (File file : File.values()) {
+            Square square = Square.of(file, rank);
+            Piece piece = boardOutput.get(square);
+
+            String pieceName = PieceName.findName(piece);
+            stringBuilder.append(pieceName);
+        }
+
+        return stringBuilder.toString();
     }
 }
